modify Pid creation
Since the most common thing to do with a PID file is check its existence
and error out if it is there, I am putting this functionality into the
create call.

+++ Pid.py
@@ -1,4 +1,5 @@
 import os
+import sys
 
 # Functions to help with PID managmnet, we will use PID files to ensure only
 # one instance of the tests are running at a time.
@@ -6,6 +7,10 @@ import os
 pid_file_path = '/var/run/zfs_perf_test.pid'
 
 def create_pid_file():
+    if pid_file_exists():
+        print("Pid file exists, please ensure that no other instances of " +
+        "these performance tests are running.")
+        sys.exit(1)
     pid = os.getpid()
     pid_file_object = open(pid_file_path, 'w')
     pid_file_object.write(str(pid))
